Crate snowbridge_runtime_primitives

Structs

UncheckedExtrinsic

Simple blob to hold an extrinsic without committing to its format and ensure it is serialized correctly.

Type Definitions

AccountId

Some way of identifying an account on the chain. We intentionally make it equivalent to the public key of our transaction signing scheme.

AccountIndex

The type for looking up accounts. We don’t expect more than 4 billion of them, but you never know…

Address

The address format for describing accounts.

Balance

Balance of an account.

Block

Opaque block

BlockId

Opaque block id

BlockNumber

An index to a block.

DigestItem

Digest item type.

Hash

A hash of some data used by the chain.

Header

Opaque header

Index

Index of a transaction in the chain.

Signature

Alias to 512-bit hash when used in the context of a transaction signature on the chain.